Refactoring. Using tendermint node client interface
